+++ Core/Launchpad/sys_text.py
@@ -1,8 +1,8 @@
 # Desc: Text processing shell commands - RPCortex Nebula OS
 # File: /Core/Launchpad/sys_text.py
-# Last Updated: 4/1/2026
+# Last Updated: 6/9/2026
 # Lang: MicroPython, English
-# Version: v0.8.1
+# Version: v0.8.2
 # Author: dash1101
 #
 # Provides grep, wc, find, sort, uniq, hex, basename, dirname.